+++ backend/src/baserow/contrib/database/fields/handler.py
@@ -6,6 +6,7 @@ from django.conf import settings
 
 from baserow.core.exceptions import UserNotInGroupError
 from baserow.core.utils import extract_allowed, set_allowed_attrs
+from baserow.contrib.database.db.schema import lenient_schema_editor
 
 from .exceptions import (
     PrimaryFieldAlreadyExists, CannotDeletePrimaryField, CannotChangeFieldType
@@ -109,7 +110,10 @@ class FieldHandler:
 
         # Change the field in the table schema.
         connection = connections[settings.USER_TABLE_DATABASE]
-        with connection.schema_editor() as schema_editor:
+        with lenient_schema_editor(
+            connection,
+            field_type.get_alter_column_type_function(connection, field)
+        ) as schema_editor:
             to_model = field.table.get_model(field_ids=[], fields=[field])
             from_model_field = from_model._meta.get_field(field.db_column)
             to_model_field = to_model._meta.get_field(field.db_column)