+++ changelog.md
@@ -10,3 +10,5 @@
 * The next field is now selected when the tab character is pressed when a field is
   selected.
 * Changed the styling of the notification alerts.
+* Fixed failing when changing field type and the a data value wasn't in the correct
+  format.